+++ features/borrow/manage/pipes/manageVault.ts
@@ -34,8 +34,7 @@ import { VaultErrorMessage } from '../../../form/errorMessagesHandler'
 import { VaultWarningMessage } from '../../../form/warningMessagesHandler'
 import { BalanceInfo, balanceInfoChange$ } from '../../../shared/balanceInfo'
 import { BaseManageVaultStage } from '../../../types/vaults/BaseManageVaultStage'
-import { VaultHistoryEvent } from '../../../vaultHistory/vaultHistory'
-import { createHistoryChange$ } from './manageHistory'
+import { createHistoryChange$, VaultHistoryEvent } from '../../../vaultHistory/vaultHistory'
 import { validateErrors, validateWarnings } from './manageVaultValidations'
 import { BorrowManageVaultViewStateProviderInterface } from './viewStateProviders/borrowManageVaultViewStateProviderInterface'
 import { ManageVaultAllowanceChange } from './viewStateTransforms/manageVaultAllowances'
