+++ helpers/stories/ManageMultiplyVaultStory.tsx
@@ -88,7 +88,6 @@ export function manageMultiplyVaultStory({
 
     const ctx = ({
       vaultHistory$: memoize(() => of([])),
-      vaultMultiplyHistory$: memoize(() => of([])),
       context$: of({ etherscan: 'url' }),
       generalManageVault$: memoize(() =>
         createGeneralManageVault$(
