+++ helpers/stories/ManageVaultStory.tsx
@@ -92,7 +92,6 @@ export function manageVaultStory({
     }, [])
 
     const ctx = ({
-      vaultMultiplyHistory$: memoize(() => of([])),
       vaultHistory$: memoize(() => of([])),
       context$: of({ etherscan: 'url' }),
       generalManageVault$: memoize(() =>
