+++ public/locales/en/common.json
@@ -210,7 +210,7 @@
     "close_vault_to_collateral": "Closed Vault, exited to collateral",
     "sold": "Sold",
     "total-fees": "Total fees",
-    "stoploss": "Stop Loss trigger {{eventType}}"
+    "stop-loss": "Stop Loss trigger {{eventType}}"
   },
   "jwt-auth-failed": "Signature failed",
   "jwt-auth-in-progress": "Waiting for signature...",
